requirements.txt
beautifulsoup4==4.9.3
blis==0.7.4
catalogue==2.0.4
cdifflib==1.2.5
certifi==2021.5.30
charset-normalizer==2.0.3
click==7.1.2
cssselect==1.1.0
cymem==2.0.5
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.1.0/en_core_web_sm-3.1.0-py3-none-any.whl
feedfinder2==0.0.4
feedparser==6.0.8
filelock==3.0.12
idna==3.2
importlib-metadata==4.6.1
jieba3k==0.35.1
Jinja2==3.0.1
joblib==1.0.1
lxml==4.6.3
MarkupSafe==2.0.1
murmurhash==1.0.5
newspaper3k==0.2.8
nltk==3.6.2
numpy==1.21.1
oauthlib==3.1.1
packaging==21.0
pathy==0.6.0
Pillow==8.3.1
preshed==3.0.5
pydantic==1.8.2
pyparsing==2.4.7
PySocks==1.7.1
python-dateutil==2.8.2
PyYAML==5.4.1
regex==2021.7.6
requests==2.26.0
requests-file==1.5.1
requests-oauthlib==1.3.0
sgmllib3k==1.0.0
six==1.16.0
smart-open==5.1.0
soupsieve==2.2.1
spacy==3.1.1
spacy-legacy==3.0.8
srsly==2.4.1
thinc==8.0.8
tinysegmenter==0.3
tldextract==3.1.0
tqdm==4.61.2
tweepy==3.10.0
typer==0.3.2
typing-extensions==3.10.0.0
urllib3==1.26.6
wasabi==0.8.2
zipp==3.5.0
